refactor(subway): rename functions, classes
- 함수명 변경
- 클래스명 변경
