+++ src/main/java/subway/service/abstraction/validation/NameAddingValidationInterface.java
@@ -4,5 +4,5 @@ public interface NameAddingValidationInterface {
     boolean checkNameDuplication(String name);
     boolean checkNameLength(String name);
     boolean checkNameLastCharacter(String name);
-    boolean checkValidation(String name);
+    boolean checkAddingValidation(String name);
 }
