+++ src/main/java/subway/view/output/station/StationInformationView.java
@@ -1,8 +1,8 @@
-package subway.view.output;
+package subway.view.output.station;
 
 import subway.type.InformationType;
 
-public class InformationView {
+public class StationInformationView {
     public static void printStationAddingInformation() {
         System.out.println();
         System.out.println(InformationType.STATION_ADDING_INFORMATION.getInformation());
